# Optimization batch 12: miscellaneous unthemed stuff